[PATCH] Introduce an extra runstate for monitoring thread startup
This way we can detect whether a thread is hanging in init, and take
it down after a timeout. This would have caught the /dev/random
problems fixed yesterday.

Also limit the Client message printed at the beginnig, to only show
the first and last of identical job types instead of everything
in between.