Add support to display slave device stats for software RAIDs.
When running tests on a device created using mdadm, the relevant
disk statistics must also come from the individual
disks that constitute the RAID, because the software RAID device
is just a virtual device.

This patch goes through the 'slaves' directory in the software
RAID device's directory in /sys/block, gets a list of all the
slave devices and creates a master-slave relationship between
the virtual device and the actual disks.

The individual slave disks' stats are displayed along with the
master's disk stats at the end of the test.

Signed-off-by: Shehjar Tikoo <shehjart@gelato.unsw.edu.au>
Signed-off-by: Jens Axboe <jens.axboe@oracle.com>