Fixup iolog replay and add 'wait' command
I have been working with Mike O'Sullivan, a senior lecturer at the University
of Auckland, New Zealand. He has been adapting fio 1.34.1 to work with iolog
files and also to be able to read a "wait" command to allow the iolog file to
specify a wait (in millisecs) before the next io. Mike's not completely
confident that his changes are correct in the overall fio picture, but they are
working with the iolog files we are using. What we did:

1) We incorporated wait into iolog file reading;
2) We made some modifications so that read_iolog worked.

We have attached a patch for fio-1.34.1.tar.bz2 along with some demonstration
fio config files and iolog files.

We are doing this so we can use an open source SPC-1 emulator to
generate SPC-1 like workloads and then run them in fio.

Signed-off-by: Jens Axboe <jens.axboe@oracle.com>