Document that write_iolog is unsafe for concurrent jobs
The write_iolog parameter should not be used in the global section since
each job will independently write to the same file.  The output will be
interspersed and the file will be corrupt.

Document that each job should write to its own iolog file.  This will
save users time when trying out the iolog feature for the first time.

Signed-off-by: Stefan Hajnoczi <stefanha@linux.vnet.ibm.com>

Updated by me to make the same change to the man page.

Signed-off-by: Jens Axboe <jaxboe@fusionio.com>