gfio: get rid of invisible data
At first the invisible data seemed like an elegant way to get
reasonable graph ticks when no data was really present, but it
proved to be an ugly hack that caused other problems later,e.g.
when you want to cap other data so the xy graph scrolls, the
invisible data kept it from scrolling, etc.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>