gfio: add per-job option view
The option view is now a master notebook of jobs that got added
from a given job file, with each tab holding a new notebook
with tabs for the options.

This is pretty slow for big jobs. Adding eg the 1mbs-clients.fio
example job file shows this nicely, since it contains 32 jobs.
Will need to look into how to speed this up.

Signed-off-by: Jens Axboe <axboe@kernel.dk>