gfio: change job option view
Instead of having a top notebook containing jobs, and subnotebooks
containing options, have just the one subnotebook and change the
viewed options by selecting the job in a combo box at the top.
This greatly reduces the number of widgets displayed for a big
number of jobs, speeding it up.

Signed-off-by: Jens Axboe <axboe@kernel.dk>