server: move fio_net_cmd to the end of fio_net_int_cmd
fio_net_cmd is variably sized. We don't use any transfer space
for the internal command, but the compiler might complain about
this (since it has no way to know this fact). Move fio_net_cmd
to the end of the internal command to both avoid a warning on
clang and to make this more idiot proof.

Signed-off-by: Jens Axboe <axboe@kernel.dk>